Establish validation workflow

+++ src/features/create_xy.py
@@ -12,13 +12,33 @@ from tqdm import tqdm
 from skimage.util import img_as_ubyte
 import json
 
+def confirm_ard_alignment(v_train, local_dir, df_sample):
+    '''
+    TBD if needed --
+    Cross references ARD generation survey w/ sample survey
+    to ensure plot naming convention is executed properly
+    
+    Groups sample csv file and asserts there are 196
+    rows per plot. Checks that each plot_fname in the
+    plot csv exists in the sample csv
+    '''
+    df_plots = pd.read_csv(f'{local_dir}collect_earth_surveys/plantations-train-{v_train}/ceo-plantations-train-{v_train}-plot.csv')
+    # plot_fnames = np.unique(df_plots.PLOT_FNAME)
+    # sample_fnames = np.unique(df_sample.PLOT_FNAME)
+    # assert (plot_fnames == sample_fnames).all(), print(f"Differences: {np.setxor1d(plot_fnames, sample_fnames)}")
+    plot_counts = df_sample.groupby('PLOT_FNAME').size()
+    for plot_fname, count in plot_counts.items():
+        assert count == 196, f"Error: plot_fname {plot_fname} does not have 196 rows (has {count} rows)"
+    print("All fname checks passed!")
+
 def load_ceo_csv(v_train_data, local_dir):
     '''
     Cleans up the CEO sample survey in order to create label
     arrays by creating a plantation encoding. Ensures
     all input csvs have same format. 
     Creates the plot_fname using the same naming convention
-    as the feature extraction pipeline.
+    as ARD generation pipeline, except for the **SAMPLE** scale
+    CEO survey.
     '''
 
     csv = f"{local_dir}ceo-plantations-train-{v_train_data}.csv"
@@ -29,6 +49,7 @@ def load_ceo_csv(v_train_data, local_dir):
     df.columns = [x.upper() for x in df.columns]
     df.columns = ['PLOT_ID' if x == 'PLOTID' else x for x in df.columns]
     df.columns = ['SAMPLE_ID' if x == 'SAMPLEID' else x for x in df.columns]
+
     df = df[['PLOT_ID', 
             'SAMPLE_ID', 
             'LON', 'LAT',
@@ -38,20 +59,23 @@ def load_ceo_csv(v_train_data, local_dir):
                                         'Agroforestry': 2,
                                         'Unknown': 255})
     
-    # create a col called plot_fname to hold new plot ids
+    # cross references naming convention in plot csv to ensure
+    # alignment between ard files
     df['PLOT_FNAME'] = '0'
+    plot_ids = []
+    counter = 0
+    for index, row in df.iterrows():
+        if row['PLOT_ID'] not in plot_ids:
+            plot_ids.append(row['PLOT_ID'])
+            counter += 1
+        
+        df.loc[index, 'PLOT_FNAME'] = f"{str(v_train_data[1:]).zfill(2)}{str(counter).zfill(3)}"
 
-    if len(df) > 0:
-        plot_ids = []
-        counter = 0
-        for index, row in df.iterrows():
-            if row['PLOT_ID'] not in plot_ids:
-                plot_ids.append(row['PLOT_ID'])
-                counter += 1
-            # create plot fnames based on training batch (v21 will be 21)
-            df.loc[index, 'PLOT_FNAME'] = f"{str(v_train_data[1:]).zfill(2)}{str(counter).zfill(3)}"
     print("Writing clean csv...")
     df.to_csv(csv)
+
+    confirm_ard_alignment(v_train_data, local_dir, df)
+
     return df
 
 
@@ -80,6 +104,8 @@ def reconstruct_images(plot, df):
 
 def create_label_arrays(v_train_data, local_dir):
     '''
+    Checks if label arrays already exist on file.
+    Assumption is that they don't need to be recreated.
     Iterates through each training data batch and 
     ensures csvs are in the correct format with 
     load_ceo_csv, then gathers the plot ids and
@@ -89,15 +115,19 @@ def create_label_arrays(v_train_data, local_dir):
     directory = f"{local_dir}train-labels/"
     
     for i in v_train_data:
-        print(f"Creating label arrays for {i}")
-        df = load_ceo_csv(i, local_dir)
-        plot_ids = sorted(df['PLOT_ID'].unique())
-        plot_fname = sorted(df['PLOT_FNAME'].unique())        
-        for i, x in zip(plot_ids, plot_fname):
-            #print(f"plot_ids:{i} plot_fname: {x}")
-            plot = reconstruct_images(i, df)
-            plot = np.array(plot)
-            np.save(f"{directory}{str(x).zfill(5)}.npy", plot)
+        labels_exist = any(file.startswith(i[1:]) for file in os.listdir(directory))
+        if labels_exist:
+            print(f"Label arrays exist for {i}, skipping")
+        else:
+            print(f"Creating label arrays for {i}")
+            df = load_ceo_csv(i, local_dir)
+            plot_ids = sorted(df['PLOT_ID'].unique())
+            plot_fname = sorted(df['PLOT_FNAME'].unique())        
+            for i, x in zip(plot_ids, plot_fname):
+                #print(f"plot_ids:{i} plot_fname: {x}")
+                plot = reconstruct_images(i, df)
+                plot = np.array(plot)
+                np.save(f"{directory}{str(x).zfill(5)}.npy", plot)
     
     return plot
 
@@ -347,7 +377,8 @@ def build_training_sample(train_batch, classes, params_path, logger):
     Gathers training data plots from collect earth surveys (v1, v2, v3, etc)
     and loads data to create a sample for each plot. Removes ids where there is no
     cloud-free imagery or "unknown" labels.
-    Creates new label files according to params.
+    Create labels should be True any time a new CEO survey is added. This triggers
+    csv cleaning and creates label arrays for the new training batch.
 
     Combines samples as X and loads labels as y for input to the model.
     """
@@ -356,10 +387,10 @@ def build_training_sample(train_batch, classes, params_path, logger):
 
     train_data_dir = params["data_load"]["local_prefix"]
     ttc_feats_dir = params["data_load"]["ttc_feats_dir"]
-    plot_ids = gather_plot_ids(train_batch, train_data_dir, ttc_feats_dir, logger)
-    logger.info(f"{len(plot_ids)} plots will be used in training.")
     if params['data_load']['create_labels']:
         create_label_arrays(train_batch, train_data_dir)
+    plot_ids = gather_plot_ids(train_batch, train_data_dir, ttc_feats_dir, logger)
+    logger.info(f"{len(plot_ids)} plots will be used in training.")
 
     # create empty x and y array based on number of plots
     # x.shape is (plots, 14, 14, n_feats) y.shape is (plots, 14, 14)
